Anul_IV/PAD/Lab_3_Proxy_Application/requirements.txt
Flask==2.2.3
redis==4.5.1
requests==2.31.0
cassandra-driver==3.25.0